requirements.txt
streamlit
pandas
plotly
requests
fpdf2
